+++ lib/Models/Catalog/Ows/callWebCoverageService.ts
@@ -1,28 +1,28 @@
 "use strict";
 
-import URI from "urijs";
 import i18next from "i18next";
 import { runInAction } from "mobx";
-import loadBlob from "../../../Core/loadBlob";
 import CesiumMath from "terriajs-cesium/Source/Core/Math";
 import Rectangle from "terriajs-cesium/Source/Core/Rectangle";
 import RequestErrorEvent from "terriajs-cesium/Source/Core/RequestErrorEvent";
-import CommonStrata from "../../Definition/CommonStrata";
-import createStratumInstance from "../../Definition/createStratumInstance";
-import proxyCatalogItemUrl from "../proxyCatalogItemUrl";
-import ResultPendingCatalogItem from "../ResultPendingCatalogItem";
-import UserDrawing from "../../UserDrawing";
-import WebMapServiceCatalogItem, {
-  formatDimensionsForOws
-} from "./WebMapServiceCatalogItem";
+import URI from "urijs";
+import filterOutUndefined from "../../../Core/filterOutUndefined";
+import flattenNested from "../../../Core/flattenNested";
 import isDefined from "../../../Core/isDefined";
-import makeRealPromise from "../../../Core/makeRealPromise";
+import loadBlob from "../../../Core/loadBlob";
+import Result from "../../../Core/Result";
 import TerriaError from "../../../Core/TerriaError";
 import { InfoSectionTraits } from "../../../Traits/TraitsClasses/CatalogMemberTraits";
+import CommonStrata from "../../Definition/CommonStrata";
+import createStratumInstance from "../../Definition/createStratumInstance";
+import UserDrawing from "../../UserDrawing";
+import proxyCatalogItemUrl from "../proxyCatalogItemUrl";
+import ResultPendingCatalogItem from "../ResultPendingCatalogItem";
+import WebMapServiceCatalogItem from "./WebMapServiceCatalogItem";
 
-var sprintf = require("terriajs-cesium/Source/ThirdParty/sprintf").default;
+const sprintf = require("terriajs-cesium/Source/ThirdParty/sprintf").default;
 
-export const callWebCoverageService = function(
+export function callWebCoverageService(
   wmsCatalogItem: WebMapServiceCatalogItem
 ): Promise<undefined | { name: string; file: Blob }> {
   return new Promise((resolve, reject) => {
@@ -44,7 +44,7 @@ export const callWebCoverageService = function(
             );
         }
       },
-      onCleanUp: () => {
+      onCleanUp: async () => {
         if (isDefined(rectangle)) {
           launch(wmsCatalogItem, rectangle)
             .then(resolve)
@@ -59,38 +59,104 @@ export const callWebCoverageService = function(
 
     userDrawing.enterDrawMode();
   });
-};
+}
 
 async function launch(
   wmsCatalogItem: WebMapServiceCatalogItem,
   bbox: Rectangle
+) {
+  if (!wmsCatalogItem.linkedWcsUrl || !wmsCatalogItem.linkedWcsCoverage) return;
+
+  const url = getCoverageUrl(wmsCatalogItem, bbox).raiseError(
+    wmsCatalogItem.terria,
+    `Error occurred while generating WCS GetCoverage URL`
+  );
+
+  if (url) {
+    return callUrl(wmsCatalogItem, url);
+  }
+}
+
+function getCoverageUrl(
+  wmsCatalogItem: WebMapServiceCatalogItem,
+  bbox: Rectangle
+): Result<string | undefined> {
+  try {
+    let error: TerriaError | undefined = undefined;
+
+    // Get dimensionSubsets
+    const dimensionSubsets: string[] = [];
+    if (wmsCatalogItem.dimensions) {
+      Object.entries(wmsCatalogItem.dimensions).forEach(([key, values]) => {
+        if (isDefined(values)) {
+          // If we have multiple values for a particular dimension, they will be comma separated
+          // WCS only supports a single value per dimension - so we take the first value
+          const valuesArray = values.split(",");
+          const value = valuesArray[0];
+
+          if (valuesArray.length > 1) {
+            error = new TerriaError({
+              title: "Warning: export may not reflect displayed data",
+              message: `WebCoverageService (WCS) only supports one value per dimension.\n\n  Multiple dimension values have been set for \`${key}\`. WCS GetCoverage request will use the first value (\`${key} = "${value}"\`).`,
+              importance: 1
+            });
+          }
+
+          // Wrap string values in double quotes
+          dimensionSubsets.push(
+            `${key}(${typeof value === "string" ? `"${value}"` : value})`
+          );
+        }
+      });
+    }
+
+    // Make query parameter object
+
+    const query = {
+      service: "WCS",
+      request: "GetCoverage",
+      version: "2.0.0",
+      coverageId: wmsCatalogItem.linkedWcsCoverage,
+      format: wmsCatalogItem.linkedWcsParameters.outputFormat,
+
+      // Add subsets for bbox, time and dimensions
+      subset: filterOutUndefined(
+        flattenNested([
+          `Long(${CesiumMath.toDegrees(bbox.west)},${CesiumMath.toDegrees(
+            bbox.east
+          )})`,
+          `Lat(${CesiumMath.toDegrees(bbox.south)},${CesiumMath.toDegrees(
+            bbox.north
+          )})`,
+          // Add time dimension
+          wmsCatalogItem.currentDiscreteTimeTag
+            ? `Time("${wmsCatalogItem.currentDiscreteTimeTag}")`
+            : undefined,
+          // Add other dimensions
+          ...dimensionSubsets
+        ])
+      ),
+
+      subsettingCrs: "EPSG:4326",
+      outputCrs: wmsCatalogItem.linkedWcsParameters.outputCrs,
+      styles: wmsCatalogItem.styles
+    };
+
+    return new Result(
+      new URI(wmsCatalogItem.linkedWcsUrl).query(query).toString(),
+      error
+    );
+  } catch (e) {
+    return Result.error(e);
+  }
+}
+
+async function callUrl(
+  wmsCatalogItem: WebMapServiceCatalogItem,
+  url: string
 ): Promise<{ name: string; file: Blob }> {
-  bbox.west = CesiumMath.toDegrees(bbox.west);
-  bbox.south = CesiumMath.toDegrees(bbox.south);
-  bbox.east = CesiumMath.toDegrees(bbox.east);
-  bbox.north = CesiumMath.toDegrees(bbox.north);
-
-  const query: any = {
-    service: "WCS",
-    request: "GetCoverage",
-    version: "1.0.0",
-    format: "GeoTIFF",
-    crs: "EPSG:4326",
-    width: 1024,
-    height: Math.round((1024 * bbox.height) / bbox.width),
-    coverage: wmsCatalogItem.linkedWcsCoverage,
-    bbox: `${bbox.west},${bbox.south},${bbox.east},${bbox.north}`,
-    ...formatDimensionsForOws(wmsCatalogItem.dimensions),
-    time: wmsCatalogItem.currentDiscreteTimeTag,
-    styles: wmsCatalogItem.styles
-  };
-
-  var uri = new URI(wmsCatalogItem.linkedWcsUrl).query(query);
-
-  var url = proxyCatalogItemUrl(wmsCatalogItem, uri.toString());
-
-  var now = new Date();
-  var timestamp = sprintf(
+  const now = new Date();
+  const timestamp = sprintf(
     "%04d-%02d-%02dT%02d:%02d:%02d",
     now.getFullYear(),
     now.getMonth() + 1,
@@ -100,13 +166,13 @@ async function launch(
     now.getSeconds()
   );
 
-  const asyncResult = new ResultPendingCatalogItem(
+  const pendingWorkbenchItem = new ResultPendingCatalogItem(
     `WCS: ${wmsCatalogItem.name} ${timestamp}`,
     wmsCatalogItem.terria
   );
 
   runInAction(() => {
-    asyncResult.setTrait(
+    pendingWorkbenchItem.setTrait(
       CommonStrata.user,
       "shortReport",
       i18next.t("models.wcs.asyncPendingDescription", {
@@ -115,38 +181,28 @@ async function launch(
       })
     );
 
+    // Create info section from URL query parameters
     const info = createStratumInstance(InfoSectionTraits, {
       name: "Inputs",
-      content:
-        '<table class="cesium-infoBox-defaultTable">' +
-        [
-          { name: "Bounding box", value: query.bbox },
-          { name: "Format", value: query.format }
-        ].reduce(function(previousValue, parameter) {
-          return (
-            previousValue +
-            "<tr>" +
-            '<td style="vertical-align: middle">' +
-            parameter.name +
-            "</td>" +
-            "<td>" +
-            parameter.value +
-            "</td>" +
-            "</tr>"
-          );
-        }, "") +
-        "</table>"
+      content: `<table class="cesium-infoBox-defaultTable">${Object.entries(
+        new URI(url).query(true)
+      ).reduce<string>(
+        (previousValue, [key, value]) =>
+          `${previousValue}<tr><td style="vertical-align: middle">${key}</td><td>${value}</td></tr>`,
+        ""
+      )}</table>`
     });
 
-    asyncResult.setTrait(CommonStrata.user, "info", [info]);
+    pendingWorkbenchItem.setTrait(CommonStrata.user, "info", [info]);
   });
 
-  asyncResult.terria.workbench.add(asyncResult);
-
+  pendingWorkbenchItem.terria.workbench.add(pendingWorkbenchItem);
   try {
-    const blob = await makeRealPromise<Blob>(loadBlob(url));
+    const blob = await loadBlob(proxyCatalogItemUrl(wmsCatalogItem, url));
 
-    runInAction(() => asyncResult.terria.workbench.remove(asyncResult));
+    runInAction(() =>
+      pendingWorkbenchItem.terria.workbench.remove(pendingWorkbenchItem)
+    );
 
     return { name: `${wmsCatalogItem.name} clip.tiff`, file: blob };
   } catch (error) {
@@ -166,9 +222,13 @@ async function launch(
           "text/xml"
         );
 
-        if (xml.documentElement.localName === "ServiceExceptionReport") {
-          const message = xml.getElementsByTagName("ServiceException")?.[0]
-            .innerHTML;
+        if (
+          xml.documentElement.localName === "ServiceExceptionReport" ||
+          xml.documentElement.localName === "ExceptionReport"
+        ) {
+          const message =
+            xml.getElementsByTagName("ServiceException")?.[0]?.innerHTML ??
+            xml.getElementsByTagName("ows:ExceptionText")?.[0]?.innerHTML;
           if (isDefined(message)) {
             error = message;
           }
@@ -187,6 +247,8 @@ async function launch(
       })
     });
   } finally {
-    runInAction(() => asyncResult.terria.workbench.remove(asyncResult));
+    runInAction(() =>
+      pendingWorkbenchItem.terria.workbench.remove(pendingWorkbenchItem)
+    );
   }
 }